Reorganize and simplify
